refactor CLI into frontend submodule, clean up command branching code
